fix a hidden bug which will surface when the first level is great than 16 bit. see the test data.

+++ src/dacs.c
@@ -148,7 +148,7 @@ FTRep* createFT(uint *list,uint listLength){
 	kvalues = optimizationk(list,listLength,&nkvalues);
 	
 	
-	ushort kval;
+	uint kval;
 	uint oldval =0;
 	uint newval =0;
